mcp-servers/spine-analysis-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spine-analysis-mcp"
version = "1.0.0"
description = "MCP Server for deep BigQuery spine dataset analysis"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "google-cloud-bigquery>=3.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0",
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
spine-analysis-mcp = "spine_analysis_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spine_analysis_mcp"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
